refactor: move Euler vault hooks to hooks/euler/
Split _vaultLookup.ts into dedicated files under hooks/euler/:
abis.ts, client.ts, and one file per hook. Dropped three hooks that
had no callers (useVaultBalance, useVaultDebt, useNativePrice).
